An nguyen

+++ src/main/java/com/example/demo/services/MovieTypeSevice.java
@@ -17,7 +17,7 @@ import java.util.List;
  */
 public interface MovieTypeSevice {
     List<MovieType> getAllMovieTypes();
-    MovieType getMovieTypeById(int id);
+    MovieType getMovieTypeById(Long id);
     MovieType saveMovieType(MovieType movieType);
-    void deleteMovieTypeById(int id);
+    void deleteMovieTypeById(Long id);
 }
